Fix critical memory bug + small cleanup
The bug was calling shrink_to_fit after getting the capacity in the
from_vec call, which was erroneously left from a previous refactor.

The with_capacity constructor was removed becuase it's misleading since
it doen't actually set any capacity. In fact DataBuffer is agnostic to
the capacity of the underlying vector.

from_slice now requires the type to implement Clone.

Finally a helpful error message was added for unsupported numeric types
when pretty printing the DataBuffer.