Fix a bug with empty buffers
This forced me to change the implementation to store element size
instead of length of the buffer, otherwise we wouldn't know the element
size if the buffer had length 0. As a result it's now more expensive to
compute the buffer length because it involves a division. The only
alternative is to store both element size and length. I decided to
sacrifice performance of len() in favor of a smaller memory footprint.