Major updates + extensions
Renamed DataBuffer to VecCopy to sound more generic as opposed to
assuming a single application.

Added VecClone which stores Clone and Drop functions for the contained
elements and manages memory using function pointers without requiring
type information.

Added a value API module to allow the users to extract a value from an
untyped Vec without needing to downcast immediately. This allows users
to move/copy and clone values in a generic way.