Bump to v0.4, remove shared-vtables feature
The feature was not additive. It is left to future work to investigate
if we ever need shared-vetables.
The current choice is to use Boxes for holding vtables in owned containers.
In contrast, using Rc prevents the containers from being used in async
contexts and using Arc has a prohibitive performance impact.