Fix an stack overflow issue due to Rust updates
At some point a new Rust update has caused some of the tests in dync to
fail in release builds.
After some investigation I believe that the issue is due to some std
algorithms placing elements of a vec on the stack which may not be able
to accomodate very large alignments. Although none of the tests use very
large alignments, it seems that the compiler pre-emptively fails even if
there could potentially be a stack overflow.