Fix overly conservative eq behaviour on Value
Value type panicked when two values with different types were compared.
This is overly restrictive because it prevents Value's of different
types to be stored in the same std collection (e.g. HashSet or Vec).
This greatly limits its applicability, and is needless. Instead of
panicking two Values of different types are now simply not equal.

This issue was not caught before because, I believe something about the
std HashTable has changed and now performs more comparisons, thus
catching this panic in another crate that uses dync.

This commit adds a test that tests Value with the HashSet to catch any
regressions in this behaviour.